refactor networking core
the read callback now properly cares about the length of the message
being read whereas before it tried to be agnostic to that and just
watched out for EOF. that doesn't work. no matter how many ways i tried
to implement it. so, next step is to refactor the code that touches
messages since i reorganized the class structure (mostly the header) as
part of this.